Drop now unused posix-source-accessor.hh header
This was fully unused now and didn't expose any public API anyway.
Drop it now, so that it can be split into unixy and windows implementation
files instead of being crammed into a single one with CPP.
