+++ src/nix/prefetch.cc
@@ -8,7 +8,7 @@
 #include "nix/expr/attr-path.hh"
 #include "nix/expr/eval-inline.hh"
 #include "nix/cmd/legacy.hh"
-#include "nix/util/posix-source-accessor.hh"
+#include "nix/util/source-accessor.hh"
 #include "nix/cmd/misc-store-flags.hh"
 #include "nix/util/environment-variables.hh"
 #include "nix/util/url.hh"
